requirements.txt
fastapi_amis_admin>=0.7.1
fastapi_user_auth>=0.7.2
fastapi-scheduler>=0.0.15
uvicorn>=0.27.1
gunicorn>=21.2.0
hypercorn>=0.16.0
loguru>=0.7.2
simplejson>=3.19.2
jinja2>=3.1.3
pysondb-v2>=2.2.0
starlette_exporter>=0.21.0
jaeger-client>=4.8.0
opentracing_instrumentation>=3.3.1
Starlette-OpenTracing>=0.1.0